Manually configure the private network interface
I am not sure of the specifics of why this happens but on vagrant 1.8.6 the network interface does not come up correctly and the private_network is attached to the `eth0` nat interface rather than the host-only interface. I tried a number of different parameters but none of them were able to configure the network appropriately. This change manually configures the static ip so that it is connected to the correct adapter. Without this change I could not access the aurora web interface.