Refactor DirectorySandbox to take the Mesos sandbox path
Rather than taking "/path/to/mesos_task_dir/sandbox" as a constructor
argument, the sandbox objects are now passed just "/path/to/mesos_task_dir"
as this simplifies a few operations that need the folder created by
Mesos.

The "sandbox" subfolder is still created with the same name, but this is
now only handled internally. There is no user-visible behaviour change.