Fix sandbox permission errors with Mesos 1.6.0
Mesos 1.6.0 creates sandboxes with permissions 750 rather than 755. This
breaks the assumption of Thermos that non-privileged processes can still
read the sandbox content.

This change is necessary even when container images are used via the
Mesos containerizer, as certain processes such as the stdout log
rotation continue to run outside of the image.